backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
statsmodels==0.14.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
firebase-admin==6.2.0
python-multipart==0.0.6
pydantic==2.7.0
pydantic-core==2.18.1
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.3
